refactor: modify page and components directory structure
게시판 카테고리 별 분류에 따른 구조 변경
기존 페이지 템플릿으로 분리 및 재사용
