+++ src/components/organisms/boardPageTemplate/CreatePost.tsx
@@ -18,7 +18,7 @@ import Preview from 'public/preview.svg';
 import PreviewCheck from 'public/preview_check.svg';
 import { InputProps } from '~/types/write';
 
-const Write = () => {
+const WritePost = ({ boardCategory }: { boardCategory: 'common' | 'info' | 'qna' }) => {
   const router = useRouter();
   const [editMode, setEditMode] = useState<typeof MODE[number]>('all');
   const [chipList, setChipList] = useState<string[]>([]);
@@ -40,11 +40,13 @@ const Write = () => {
   // const
   const onSubmit = async () => {
     // console.log(inputRef.current[WRITE_REF.title]?.value, editorContent);
-    const res = await Board.create({ content: editorContent, title: inputRef.current[WRITE_REF.title]?.value }).catch(
-      (err) => {
-        console.log(err);
-      },
-    );
+    const res = await Board.create({
+      content: editorContent,
+      title: inputRef.current[WRITE_REF.title]?.value,
+      boardCategory,
+    }).catch((err) => {
+      console.log(err);
+    });
     if (!res) router.push('/board');
   };
   return (
@@ -78,4 +80,4 @@ const Write = () => {
     </Container>
   );
 };
-export default Write;
+export default WritePost;
