+++ src/pages/board/common/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from 'next';
+
+import CommonBoard from '@components/organisms/boardPageTemplate';
+
+const CommonBoardPage: NextPage = () => {
+  return <CommonBoard boardCategory="common" />;
+};
+export default CommonBoardPage;
