+++ src/pages/board/info/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from 'next';
+
+import InfoBoard from '@components/organisms/boardPageTemplate';
+
+const InfoBoardPage: NextPage = () => {
+  return <InfoBoard boardCategory="info" />;
+};
+export default InfoBoardPage;
