+++ src/pages/board/info/write.tsx
@@ -1,0 +1,6 @@
+import WritePostTemplate from '@components/organisms/boardPageTemplate/CreatePost';
+
+const WritePost = () => {
+  return <WritePostTemplate boardCategory="info" />;
+};
+export default WritePost;
